# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XR5_0TrainingRepo",
    "version": "1.0"
  },
  "paths": {
...
    "/xr50/training-repo/xr50app-management/XR50App/{appName}": {
      "get": {
        "tags": [
          "XR50App"
        ],
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/XR50App"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XR50App"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/XR50App"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "XR50App"
        ],
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "XR50App": {
        "type": "object",
        "properties": {
          "owncloudGroup": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "owncloudDirectory": {
            "type": "string",
            "nullable": true
          },
          "appName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}